backend/requirements.txt
Flask
PyMuPDF
python-docx
pdfplumber
scikit-learn
textstat
language-tool-python